refactor: remove state ful widget and use getx controllers instead
