+++ lib/views/compose/widgets/from_avatar_view.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:ndk/ndk.dart';
+import 'package:nostr_mail_client/controllers/auth_controller.dart';
+import 'package:nostr_mail_client/models/from_option.dart';
+import 'package:nostr_mail_client/widgets/nostr_avatar.dart';
+
+class FromAvatarView extends StatelessWidget {
+  final FromOption option;
+
+  const FromAvatarView({super.key, required this.option});
+
+  @override
+  Widget build(BuildContext context) {
+    final authController = Get.find<AuthController>();
+    final pubkey = authController.publicKey!;
+
+    return NostrAvatar(
+      pubkey: pubkey,
+      metadata: Metadata(
+        pubKey: pubkey,
+        picture: option.picture,
+        displayName: option.displayName,
+      ),
+      radius: 14,
+    );
+  }
+}
